requirements-test.txt
pytest>=8.0.0
pytest-cov>=4.0.0
pytest-mock>=3.12.0
pytest-asyncio>=0.23.0
click>=8.0.0
newsapi-python>=0.2.7
requests>=2.25.0
ruff>=0.1.0